requirements.txt
scikit-image
jupyter
numpy
scipy
pytest
pytest-flask
pytest-mock
rq
redis
torch
torchvision
flask
pillow==6.0.0
